+++ zebROS_ws/src/behaviors/launch/2023_balance.launch
@@ -6,9 +6,9 @@
         <!-- <param name="/use_sim_time" value="true"/> -->
         
 		<group ns="balance_position">
-            <node name="balancing_server" pkg="behaviors" type="2023_balancing_server.py" output="screen">
+            <!-- <node name="balancing_server" pkg="behaviors" type="2023_balancing_server.py" output="screen">
 
-            </node>
+            </node> -->
 
 
             <!-- Less complicated verson of hold position, because we can only move X and we use pitch-->
@@ -18,21 +18,21 @@
                 <param name="topic_from_plant" value="position_balance_pid/pitch_state_pub" /> 
                 <param name="topic_from_controller" value="position_balance_pid/x_command" /> 
                 <param name="command_timeout" value="0.25" />
-                <param name="enable_topic" value="balance_server_pid/pid_enable" />
+                <param name="enable_topic" value="x_position_balance_pid/pid_enable" /> <!-- make sure this is the same as below, same for other topics -->
                 <param name="name" value="balance_server_pid" />
                 <param name="imu_sub_topic" value="/imu/zeroed_imu" /> 
                 <param name="angle_threshold" value="0.0174533"/>  <!-- 1 degree -->
-
+                
             </node>
 
             <node name="x_position_balance_pid" pkg="pid" type="controller" output="screen" >
                 <param name="node_name" value="x_position_balance_pid" />
                 <param name="Kp" value="-0.5" />
                 <param name="Ki" value="0.0" />
-                <param name="Kd" value="-0.1" /> 
-                <param name="upper_limit" value="10" />
-                <param name="lower_limit" value="-10" />
-                <param name="windup_limit" value="10" />
+                <param name="Kd" value="0.0" /> 
+                <param name="upper_limit" value="1" />
+                <param name="lower_limit" value="-1" />
+                <param name="windup_limit" value="0.1" />
                 <param name="max_loop_frequency" value="100.0" />
                 <param name="min_loop_frequency" value="100.0" />
