+++ zebROS_ws/src/talon_swerve_drive_controller/config/2022_twist_mux.yaml
@@ -29,5 +29,9 @@ topics:
   topic   : /hold_distance/hold_position_pid/swerve_drive_controller/cmd_vel
   timeout : 0.2
   priority: 130
-
+-
+  name    : balancing
+  topic   : /balance_position/balancer_server/swerve_drive_controller/cmd_vel
+  timeout : 0.2
+  priority: 130
 locks: []
